Add support for `<select>`
