+++ lib/handlers/index.js
@@ -24,6 +24,7 @@ var paragraph = require('./paragraph')
 var quote = require('./q')
 var root = require('./root')
 var row = require('./table-row')
+var select = require('./select')
 var strong = require('./strong')
 var table = require('./table')
 var text = require('./text')
@@ -66,7 +67,6 @@ exports.optgroup = ignore
 exports.option = ignore
 exports.param = ignore
 exports.script = ignore
-exports.select = ignore
 exports.shadow = ignore
 exports.source = ignore
 exports.spacer = ignore
@@ -165,6 +165,7 @@ exports.pre = code
 exports.q = quote
 exports.s = del
 exports.samp = inlineCode
+exports.select = select
 exports.strike = del
 exports.strong = strong
 exports.summary = paragraph
