+++ test/fixtures/select-optgroup-option/index.html
@@ -17,10 +17,45 @@
 </p>
 
 <p>
-  And these:
+  This one:
+  <select multiple size="2">
+    <option>Value 1</option>
+    <option selected>Value 2</option>
+    <option>Value 3</option>
+  </select>
+</p>
+
+<p>
+  Or this:
+  <select size="2">
+    <option>Value 1</option>
+    <option selected>Value 2</option>
+    <option>Value 3</option>
+  </select>
+</p>
+
+<p>
+  Or this:
+  <select size="2">
+    <option>Value 1</option>
+    <option>Value 2</option>
+    <option>Value 3</option>
+  </select>
+</p>
+
+<p>
+  Or this:
+  <select multiple>
+    <option>Value 1</option>
+    <option selected>Value 2</option>
+    <option selected>Value 3</option>
+  </select>
+</p>
+
+<p>
+  Or this:
   <select>
-    <option value="value1">Value 1</option>
-    <option value="value2" selected>Value 2</option>
-    <option value="value3">Value 3</option>
+    <option disabled>Value 1</option>
+    <option disabled selected>Value 2</option>
   </select>
 </p>
